RagWithScraper/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain_community
faiss-cpu
beautifulsoup4
requests
python-dotenv
pydantic
tiktoken
numpy
tqdm
pytest
httpx
load_dotenv
requests
firecrawl
openai
instructor
pandas
tqdm
pydantic
pymongo
ragas
matplotlib
seaborn
uvicorn
gunicorn
sentence-transformers
scikit-learn
numpy

